Update customHeaders class.
